Cambios a sedes
La tabla sedes ahora contiene dos nuevas columnas: latitud y longitud, esto para colocar los marcadores en el mapa. La creación y edición de sedes ahora usa Geolocación de google para encontrar las coordenadas mas cercanas.

+++ api/entity/Sedes.js
@@ -15,10 +15,10 @@ module.exports = new EntitySchema({
             type: "int",
         },
         latitud:{
-            type: "int"
+            type: "decimal"
         },
         longitud:{
-            type: "int"
+            type: "decimal"
         } 
     },
     relations: {
